pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vectormancer"
version = "0.1.0"
description = "Local-first vector search for your files, notes, and code."
readme = "README.md"
authors = [{name = "Zhafira Elham"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.7",
    "pydantic>=2.6.0",
    "PyYAML>=6.0.0",
    "watchdog>=4.0.0",
    "faiss-cpu>=1.7.4; platform_system != 'Windows'",
    "pypdf>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "markdown-it-py>=3.0.0",
    "sentence-transformers>=3.0.0",
    "rich>=13.7.0",
]

[project.scripts]
vectormancer = "vectormancer.cli:main"